+++ BE/src/main/java/louie/hanse/issuetracker/domain/Member.java
@@ -23,4 +23,8 @@ public class Member {
         this.socialId = socialId;
         this.avatarImageUrl = avatarImageUrl;
     }
+
+    public Long getId() {
+        return id;
+    }
 }
